Add a Clock trait and FixedClock implementation.
This commit introduces a `Clock` trait to `oak_time` to provide an
abstraction for time sources. This allows for dependency injection of
time, which is particularly useful for testing.

A `FixedClock` implementation is also included, which always returns a
pre-configured time. This is useful for testing components that depend
on the current time, allowing for deterministic tests.

This is intended to be a replacement for the Clock type in the
oak_attestation_verification_types crate.

Change-Id: I1ffeac732b052205ee85192cb6d4227c1da3e61d
ACKNOWLEDGE_FAILING_COPYBARA_IMPORT=Guitar