Introduce a `report_attestation_token()` function for computing full verification reports for JWTs. Make `verify_attestation_token()` use its output.
I'd like feedback on this pattern; it's quite different to the one I used for signature-based attestation reports. One important thing to note is that this pattern "infects" a callstack, i.e. the layer above this - `ConfidentialSpacePolicy.verify()` - will no longer be able to call `verify_attestation_token()` once I introduce a `report()` function at that layer in the hierarchy.

Change-Id: Ib7f2a7b12b927a1a95ecd34352496d7c6ba9e4c9