Add simple tests to ensure stage0 doesn't outright break q35/microvm.
As the goal is "simple", I cobbled together a really minimal test kernel
that just prints that it booted successfully. (In the future I envision
that toy kernel also dumping out more information, like the ACPI
tables.)

My original plan was to use Restricted Kernel proper for this, but...
Restricted Kernel is too good these days. For example, if I wanted to
dump out ACPI tables and the like, that'd require a lot of work in
Restricted Kernel as the application can't just access arbitrary memory.

Another thing to implement would be a similar "boot successful" initrd
and pair it with linux kernel.

Change-Id: I6a6a696496523b7422afa8e8fc111c643914b1a5