doremint: Add OCI container endorsement check
This commit introduces the capability to verify endorsements for OCI
container images as the `image verify` command.

The comand takes an image reference and an endorsement claims file as
input. It verifies the signature on the endorsement, and that it has
been included in Rekor's transparency log.

Key changes:
- Added a crate to handle Cosign and Rekor logic.
- Included the Rekor public key for signature verification.
- Refactored the existing doremint command to be asynchronous.

ACKNOWLEDGE_FAILING_COPYBARA_IMPORT=It's not dead, it's sleeping

Bug: 430493509
Change-Id: I26ae47a82cdcf4a4a79f290274943edfb1c914af