Add attestation publisher
With this CR, your oak session configurations can now contain an
optional `AttestationPublisher` component that will be called whenever
the session receieves a new evidence set from the peer.

This component will be used as part of the client evidence logging
system we are prototyping, but may be useful for other purposes, too.

Fixed: b/409365875

Change-Id: I0331cd2ade402604ab44cb5d86f2c2e16132afbe