Make rekor::verify_rekor_log_entry() public.
In folsom, we need to verify inclusion of certain events
in rekor and we need this function.

Technically, we can use `verify_endorsement()`, however, it is not
preferred for the following:
a) We have a custom logic to check endorser signatures. We require t-of-n
instead of all signatures.
b) Validity is of no use for us right now as we don't have a good way of
accessing the current time from restricted kernel.

Bug: b/417481016
Change-Id: Iab82056e91fd2d8ee55845d772a753ca90be61e2