Load Wasm and lookup data via URLs.
This change adds support for loading in lookup data and the Wasm logic
from a URL instead of relying on files that must be linked in with the
container. Since this change relies on the `ureq` crate for sending HTTP
requests from a URI, the arguements to the binary are defined in terms
of URIs rather than URLs.

To verify, I ran
```
bazel run //oak_functions_standalone:oak_functions_standalone -- \
--wasm-uri="https://storage.googleapis.com/oak-functions-standalone-bucket/wasm/key_value_lookup.wasm" \
--lookup-data-uri="https://storage.googleapis.com/oak-functions-standalone-bucket/lookup_data/fake_weather_data.binarypb"
```

This relies on our public bucket that has both our lookup logic and
synthetic lookup data.

Bug: 424408423
Change-Id: I7e79c3d9f7407a5d9053a11dcf1d3f213c42bee6