Remove `path` flags for OakFunctions binary and update bash script.
This change removes the old `path` flags for reading files at runtime
for the Oak Functions standalone app. As a result, URIs must be passed
in as arguements to load the wasm and lookup data for the application.

The bash script is also updated to reflect the new command arguements.
The running instance can be viewed here:
https://pantheon.corp.google.com/logs/query;query=resource.type%3D%22gce_instance%22%0Aresource.labels.instance_id%3D%225844299879774729720%22;cursorTimestamp=2025-07-17T14:40:17.185305Z;startTime=2025-07-16T15:11:25.368Z?inv=1&invt=Ab3AKA&project=oak-functions-standalone

Bug: 424408423
Change-Id: I0e9c268c0f77065e4abaf129e955c0b47e94eadb