Remind developer that vhost_vsock module is needed
When running integration tests after a reboot, you probably forgot to
modprobe vhost_vsock.

If you don't want to set your system up to do this automatically, this
chnge will add a helpful little note when loading the flake environment
to let you know that you should run the command.

`kmod` is added to the base flake.nix configuration, so that CI won't
show the warning message, which would be misleading.

This is a bit easier than digging through failed integration test logs
to find that suggestion.

Change-Id: Ie87347672e7f3ba1e15f3dceebb082a0cae0f833