feat(attestation): Expose session bindings in attestation evidence
This change updates the  to include session bindings and the handshake hash.

This information was already used internally to verify that attestation evidence is bound to the current session. These changes expose the bindings to relying parties, allowing them to be stored and used as evidence in subsequent verification flows.

The changes include:
- Updating the  protobuf message.
- Modifying the Oak Session layer to populate and handle these new fields.
- Updating client implementations and tests to reflect the changes.

Bug: 419209669
Change-Id: Ia7053940611f7d703c065f93794e2d5bd25705f9