Add skeletons for the custom AddPciRootStage* table builder commands
Looks like the `AddPciHoles` custom command we dealt with previously is
marked as deprecated, and is supposed to be replaced by these two. We
haven't run into these in our VMs yet, but it looks like others have.

This is the first step: just add the command structures themselves, with
no implementation.

Bug: 434204939
Change-Id: I6a6a6964519cbf6f4f52fddb6aaaf3b33bb1195b