feat(attestation): Add OCI image verification
This commit introduces OCI image verification to the attestation flow.
It adds a new `cosign` module to `oak_attestation_gcp` to verify
cosign signatures of OCI images. It also includes the Rekor public key
for transparency log verification. In the future this could be
independent of GCP attestation.

The changes include:
- A new `cosign.rs` module for signature verification.
- The Rekor public key for transparency log verification.
- Test data for endorsements and policy claims.
- Integration of the verification logic into the attestation CLI and
  other examples.

Bug: 430493509
Change-Id: I6a5db50827922a6377847922e3ac6f45fe783fb9