Introduce new crate for freshness verification.
This introduces a new crate that relies on std to verify the
`ProofOfFreshness` proto. Currently, it only verifies that the
outputValues (between the Nist website and the proto itself) line up.

A new crate is introduced here because this must be linked into the
oak_crypto crate, which is `no-std`.

Bug: 424736845

Change-Id: Ife08f17f3ea068e665abc4b4663d2515490791cb