Refactor publish API to match latest schema
This change get the `AttestationPublisher` API up-to-date with the
pull-based `get_attestation_evidence` mechanism.

The publisher will now provide `AttestationEvidence` structs, the same
type that's returned from `get_attestation_evidence`.

Since the publisher contains bindings as well as attestation, this CR
moves the definition of the type, as well as implementation logic, out
of the `attestation` module and into the `session` module.

Bug: b/409365875

Change-Id: I137d31ca8bd96c56318a9df3e68649038eb377e1