Determine the location of the 32-bit PCI MMIO region based on VM type.
This CR has a fair amount of comments in it, contrasting how EDK2 and
SeaBIOS do things. Interestingly, for Q35, their memory maps are
completely reversed.

Next up: determine the 64-bit MMIO region. (And IO ports.) After that,
we'll start delving into the fun of actually configuring the PCI bus.

Bug: 434204939
Change-Id: I6a6a6964562758057106e5b8d49c5ea9230268dc