Reduce executor pressure in CI
* Do clippy checks as part of normal build pass
* Do format as another step in the main CI job
* Fail-fast more often

We fail fast more often because the team concluded that none of us take
advantage of thehaving multiple bits of information in the build failure
output: at most we tend to see that something failed, and then verify
the needed changes ourselves locally before re-pushing. So taking extra
executor time to run lots of jobs is not worthwhile.

Change-Id: I09a77e225b79af5c738387903f84f721e90dabed