Move tests for policies from integration to unit tests.
Mainly a preparation or invitation to add more tests. Before, we did
not have test_util so all these tests had to be integration tests.

Also change create_verifier to create_amd_verifier and fix the return
type. So the switch between the platforms will be on the client side.

Bug: 437347358
Change-Id: I265cf9b9a37182d41ea21c6e39c8b0960037ca48