Allocate resources for BARs and configure the BARs accordingly
This CR is really what the last, uh, dozen or so CRs have lead up to: we
finally allocate resources for the PCI BARs and configure the devices.

This is not the end of the road (for example, we don't touch IRQs, don't
configure PCI bridges, and don't touch PCIe/ACPI at all), but this
should fix the immediate issues we've seen.

Bug: 434204939
Change-Id: I6a6a69643a1d503471460b3bf8d30cd864a0b9e4