Keep track of the 16-bit I/O port range we will use for the PCI bus.
As always, for now, I'm hard-coding a value from SeaBIOS. This gives PCI
16k I/O ports to use and even the PCI standard recommends not relying in
I/O ports as they're a scarce commodity on x86.

While there will certainly be ways to set up a VM in a way that exhausts
these resources, my basic testing with the shapes we'd expect suggests
that's more than an order of magnitude more than we need right now.

Bug: 434204939
Change-Id: I6a6a6964718adbdfaeb2c06f4346a6651bf2c603