Create a generic allocator for all the types we'll need to track.
We now have both:
(a) windows where we can assign resources from, and
(b) can walk the BARs to find out what resources the devices need.

Unfortunately all of these are of different types. Memory can be `u32`
or `u64`; I/O ports are `u16`. But the basic allocation algorithm is the
same for all three types.

Thus, as a first step, let's define an allocator that can work with all
three types.

Bug: 434204939
Change-Id: I6a6a69647567f68ff22fcc8ed0c6cdc5fb9fe56d