Unify style of error context in oak_attestation_verification.
Concept: When providing context higher up in the stack trace, start with
a verb and use the gerund. If in doubt, just repeat the name of the
function the context is attached to, in a human readable way. Terms like
"fail" or "missing" are only used by the actual errors at the deepest
level. Exception: when catching NPEs it is fine to use context as error
(and we do that extensively).

Here is how an error with context looks:

Caused by:
    0: comparing expected values to evidence
    1: comparing kernel layer values
    2: comparing kernel image digests
    3: measurement digest RawDigest { psha2: [], sha1: [], sha2_256: [62, 124, 55, 24, 88, 242, 189, 156, 3, 40, 148, 105, 75, 213, 187, 72, 147, 242, 64, 60, 232, 160, 188, 247, 60, 7, 175, 62, 246, 163, 90, 21], sha2_512: [], sha3_512: [], sha3_384: [], sha3_256: [], sha3_224: [], sha2_384: [] } does not match any reference values)

Bug: 440516393
Change-Id: I2f29e8d4273f0915dce8e9f6aba72f5edceac202