Hard-code the available physical address bits to 40 as a workaround
We've now run into a VMM where the CPUID reports that the guest physical
address space is 48 bits, but... it's not. If we try to place the MMIO
window at the top of that address space, the VMM crashes.

I've left the existing CPUID code in there as we'll need to come back
and figure out how to see through the VMM's lies, but for now as a
temporary workaround, let's hard-code the bits to relatively safe 40.

This means if you want to start a VM with more than 800ish GiB of
memory, there will be a clash and probably bad things happen. But I
don't think we have any VMs with that size on the horizon right now.

Bug: 440475642
Bug: 434204939
Change-Id: I6a6a69644d1c8db834dd3ae4fbbb0b9e6729de06