Organize database crate into modules
The database crate contains 3 pieces with very similar names, and it can
be hard to navigate the file and tell which function you are looking at.

This cleanup divides the major components into modules, which allows for
better organization, and visibility handling.

This also fixes an issue which was that the icing unit tests in the file
were not being run, because there was no test target for the database
crate.

There is potential for further cleanup, but this is a good checkpoint.

Change-Id: Ib429230c37b6bcacd6747dd0dcef7b8055b80874