Clean-up: Move the PCI b/d/f address logic into a separate struct
The `PciAddress` was originally supposed to be exactly that, but now has
grown all kinds of accessor methods that really shouldn't be part of it.

After all, an address is an abstract concept, and should be
representable even if you won't have direct access to the PCI bus.

I've also added a bunch of tests where (a) I'll admit I used this as an
excuse to learn more about the googletest Rust implementation, and (b)
actually found the `next()` functions could panic, which is now properly
fixed.

Bug: 434204939
Change-Id: I6a6a6964f742102f2ed9baa33845c243ba7fa0c1