Change some uses of `PciAddress` to `Bdf`.
For example, the CAM read and write functions don't need `PciAddress`;
if anything, this is a circular dependency as the accessors in
`PciAddress` call the read function.

The overall goal is to refactor the `pci_read_cam` and `pci_write_cam`
functions into a trait that we could mock out for better testing
support.

Bug: 434204939
Change-Id: I6a6a696462bbaf21c11bcc8d2ddb28089066d445