Add an A2A-compatible Agent server and a corresponding Dockerfile
A2A can also be used to have a regular human-agent chat with the
server-side agent (doesn't have to be agent-to-agent). And this is
currently the most convenient approach of implementing a cloud-based
agent that a user can interact with.

Also, currently there is a startup bash script that runs the Oak Proxy
server and clients (to connect to Gemma and an MCP server). But this
will be replaced, once Oak Proxy can run both server and client in the
same binary.

Docker currently uses pip-compile to pin Python dependencies. But this
could be replaced with Bazel in future CRs: b/429197818

Bug: 432688592
Bug: 433455459
Change-Id: I3fde2f4df4196c8c868dfcac4e42bfad3a3ddf31