Introduce Metadatabase Operations and Rebasing
This introduces the concept of an operation to the icing Metadata DB.
The DB now tracks all applied operations in the instance.

It exposes a "rebase" function, which accepts a new underlying data
blob, and then creates a new instance using the provided blob as a
starting point, but re-applying all existing operations.

The "changed" and "newly_created" signals are combined into the
"applied_operations" list managed by the icing database. Now, if the
operations list is non-empty, then writeback should occur.

This will be used to implement the opportunistic concurrency writeback
strategy.

Bug: b/440413371

Change-Id: I7a0112f88f47c4faf21f72f86c43c9b8652fc041