Introduce LLM Views and enable searching within them.
This change adds `LLMView` to the `Memory` proto. The `LLMView` should
replace the `embedding` field in `Memory`. Each `LLMview` is a separate
icing document with the new `LLM_VIEW_SCHEMA`. This is required because
we want to match each llmview individually.

From now on, any search related to embedding, will search over the
LLMviews schema in icing. Text only search will still search over the
memory schema.

The LLM views is an experimental feature. Having it will allow better
exploration on semantic search.

Change-Id: If970fd32817ea5fce3921935b805f427657cc367