Make RFC3339 serialization use Z notation.
Similar to internal places, we don't want timezone information in
serialized timestamps. The Z notation works best to demonstrate
this.

Note that on deserialization we do honor timezones but discard
them immediately, as one would expect.

We also want to include milliseconds, but not nanoseconds.
Nanosecond precision is mostly to enable bijective conversion
between Instant and protobuf::Timestamp internally. It is fine
to truncate them under serialization.

Change-Id: I7787feff235dec5bdcdc8e12b89908b1d3d2c6d9