Collect all things Rekor under oak/tr/rekor.
- Remove module sigstore::rekor since it is the same as verify_rekor_log_entry
- Remove module sigstore::rekor::hashedrekord since it is the same as verify_rekor_log_entry_body

Disclaimer: oak_attestation_gcp/src/cosign.rs and
trex/doremint/src/image/verify.rs are in an intermediate state,
which works (barring lack of test coverage) but is neither nice
nor efficient. These concerns will be addressed as part of the
cited bugs. The plan is to get rid of CosignEndorsement and
CosignReferenceValues since they duplicate structures that already
exist.

Bug: 445140472
Bug: 446121597
Change-Id: Ic95b64aa9fd2c3588a2e4efaa13121d236432d03