Add Intel TDX RTMR emulator
We need a software emulation for the RTMR logic so that we can replay
the extend operations using the event log. This is needed to verify the
integrity of the event log entries in the attestation evidence.

While testing this I found a bug in the Stage 0 support for extending
RTMR 2, so I fixed that as well as part of this change.

BUG: 418687763

Change-Id: I6a6a69640c1ee9a368615f64505199bad10fb611