Update SEV-SNP secrets page to v1.58 of the spec
The secrets page version has been updated to 4. I removed the max
version check to be in line with the approach we took elsewhere to be
more permissive when parsing externally provided content.

A new field representing the mitigation vector was added to the secrets
page.

The specification no longer states that the reserved fields must be
zero, so I removed the zero-check.

See https://github.com/project-oak/oak/issues/5069 for more details.

Change-Id: I6a6a6964407a6e87705b238e23bdab4a45eba349