Introduce Read/WriteMetadataBlob and use these calls.
This introduces the new Read/WriteMetadataBlob API for the encalve app.
All usages involving metadata blobs are updated to use this API.

This will allow us to switch to the new metadata blob storing mechanism
in the host app.

A subsequent change will enable the logic needed for the opportunistic
concurrency write-back solution.

Bug: b/443329966

Change-Id: I8676c744947cde4fbe999a95cb544034b1a12163