Update oak functions standalone clients.
This change allows clients of Oak Functions Standalone to supply a
container image reference prefix to match against the container image
reference in the JWT they receive as part of Oak Session. This prefix is
provided for all current use as part of a `default` set of Oak
Functions Standalone reference values. Clients are still able to provide
their own reference values if they would like.

Bug: 439861326

Change-Id: I6c5ea4fcccfee2a4f31fdc201db84168a3c10c62