Check validity of AMD ARK, ASK and VCEK certs
We should make sure that the AMD Root Key (ARK), SEV key (ASK) and
Versioned Chip Endorsement Key (VCEK) certificates are valid at the
time when we do AMD SEV-SNP attestation verification.

Since the ASK and ARK are long lived and stable we only check validity
as part of CI for them.

Also, to make the check optional the decision whether it should be
checked is set in the reference values.

BUG: 432489444

Change-Id: I6a6a696451145d75629d329c5886e3f44e3abbb3