Create a more type-safe wrapper for `SessionConfig`
Instead of just giving a raw pointer, which can be used multiple times
(with an accompanying crash because double-free), this creates a
`unique_ptr`-inspired wrapper for the raw `SessionConfig` pointer.

As always, refactoring is hard, but eventually the raw pointer should
disappear and the `SessionConfigHolder` renamed to `SessionConfig`.

(And, we should fix the memory leak by internally using a `unique_ptr`
with a proper deallocator.)

Unfortunately there isn't a way how to make C++ compiler yell at you
when you use something after it's been moved out of, so we have the
options of (a) crash or (b) return a error absl::Status. Our current
code right now does various forms of (a), this CR changes so that it'll
do (b) instead.

I'd argue this is safer to use, but feel free to push back.

Change-Id: Ie34d070c62ec011de3717bd2e0f688eb6a6a6964