Create `CStr` versions of the attestation type constants
This makes them usable across C FFI boundaries, as Rust strings and C
strings are implemented differently.

Ideally I'd leave just the `CStr` constants around, but as always,
migration is hard so at least temporarily I've also left code in that
translates the `CStr`-s into Rust `str`-s.

Unfortunately this does mean I have to touch unsafe code, as you can't
unwrap a `Result` in a const context; thus, we have to rely on
`from_utf8_unchecked` -- but the constants _are_ valid UTF-8, so this is
not a safety issue.

Change-Id: I71c9e9b96afa4aa78ed96a505d2c91406a6a6964