Remove unused Restricted Kernel comms channels
We can no longer use or test the virtio-vsock or serial comms channels
in the restricted kernel. Both of those require the Q35 machine type
on QEMU, but the restricted kernel no longer works on Q35 machines due
to incompatible ACPI tables.

Multiple other dependencies in the repository was only used for these
comms channels, so I deleted them as well.

BUG: 463967011

Change-Id: I984ecd31841d4305a628e60182f72b126a6a6964