Add stage 0 kernel command line arg for attester.
Some of the stage 0 arguements (the Stage0DiceData C-struct and the
EventLog) are no longer necessary because of the serialized attester.
Instead, clients should rely on the serialized attester data. This
change allows clients to know where the serialized attester data is
located, preventing them from having to calculate what the address of it
is in memory beforehand.

Note that in order to compile with this new feature (which is disabled
by default), you must append `--define add_serialized_attestation_data_to_cmdline=true` to the build command.

Bug: 463325402
Change-Id: Ibf97da92e4d56c3520be31f102c2f87746eefa62