Create Oak Session (TLS) library
This introduces the Oak Session (TLS) functionality. This commit adds the
basic TLS encryption setup logic and introduces the API. Attestation
actions are not yet supported.

There will be subsequent follow-up work to improve the testing coverage
and to add additional features. This is just the first progress
checkpoint for the library.

Bug: b/448338977

Change-Id: I38c21eee7b00bbc1b4bf972147247f8d15b7b5b8