Use a single JSON file for the index instead of just a URL prefix
Define a custom JSON schema to represent an index pointing to individual endorsements, and collect some metadata there for efficient lookups.

This allows creating a CAS that is colocated with the repo hosting the index.json file, and also additional ones pointing to external location if needed in the future.

Fix various instructions to show how to endorse and verify blobs.

Bug: 462402603

Change-Id: Ic64f1810a92fabbecbc50015ca72428f6a6a6964