Address SNP cache coherency vulnerability
CVE-2025-38560 (AMD SN-3017) describes a cache coherence vulnerability
for AMD processors using SNP.

As described in the CVE:
> The specific mitigation is to touch the first and last byte of each 4K
page that is being validated.

So, let's touch the first and last byte of each PVALIDATEd page to
address this issue in software to cover stage0 use cases that do not
have the specific COHERENCY_SFW_NO CPUID bit set to 1.

We've built and ran this internally on our hosts, but I would suggest
running it on Google's internal CVMs to ensure that the 2MB pages you
folks use do not run into issues with pvalidate

Credit goes to @hvolkmer for the fix

Change-Id: I8e90b81d1069c3e3317e2094110e76796a6a6964