Introduce mcp/oak_functions_mcp folder.
The `oak_functions_mcp` folder will be where all the project code is
written to. Right now the MCP server is bare bones (and largely forked
from the existing mcp/server/ folder). This new MCP server will run on
Confidential Space and send its attestantion to the MCP Client who will
validate it and the environment it is running in.

Bug: 469747147
Change-Id: I75154a1b76daf95c14cf7ee0fe570cde671dda81